+++ .github/workflows/e2e-multiple-k8s-clusters.yaml
@@ -24,7 +24,6 @@ jobs:
       matrix:
         package:
           - changetostandalone
-          - changetoprimary
           - changetosecondary
         backup-transfer-part-size:
           - 3Mi # smaller than the PVC size
@@ -49,6 +48,9 @@ jobs:
           - package: replication
             backup-transfer-part-size: 3Mi # smaller than the PVC size
             ginkgo-flags: --label-filter="backup-failure"
+          - package: replication
+            backup-transfer-part-size: 3Mi # smaller than the PVC size
+            ginkgo-flags: --label-filter="change-to-primary"
     #runs-on: "ubuntu-22.04"
     runs-on: mantle_large_runner_16core
     timeout-minutes: 120
